renamed designerdevcomponent, button test for designing and login user issue
