+++ src/app/app.component.html
@@ -13,7 +13,7 @@
       <div class="top-menu">
         <div class="top-menu-left"></div>
         <div class="top-menu-right">
-          <a routerLink="designer" class="design-link" mat-button>{{ state.designState ? 'Continue Designing' : 'Start Designing' }}</a>
+          <a routerLink="designer" class="design-link" mat-button>{{ (appState | async)?.order ? 'Continue Designing' : 'Start Designing' }}</a>
           <ng-container *ngIf="(auth._user | async) || {} as user">
             <ng-container *ngIf="!user.uid; else loggedIn">
               <a routerLink="account/login" class="top-menu-account">
