--- src/app/designer/client/designer-client.component.css
@@ -1,199 +1,0 @@
-.pusher {
-  height: 80px;
-}
-
-.main-container {
-  width: 100%;
-  height: calc(100vh - 194px);
-  display: flex;
-  justify-content: center;
-  align-items: flex-start;
-  overflow: hidden;
-}
-
-.design-container {
-  width: 100%;
-  height: 100%;
-  background-color: rgb(48,52,58);
-  display: flex;
-  align-items: stretch;
-  overflow: hidden;
-}
-
-.designer-view {
-  flex: 1;
-  text-align: center;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  background: #eee;
-}
-
-.loading-overlay {
-  position: absolute;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  z-index: 99;
-}
-
-.loading-pdf-overlay {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  position: fixed;
-  top: 0;
-  left: 0;
-  width: 100vw;
-  height: 100vh;
-  z-index: 999999999;
-  background-color:#12c463;
-}
-
-.loading-message {
-  color: white;
-  margin-top: 8px;
-}
-
-::ng-deep .mat-progress-bar-buffer {
-  background-color: #fff;
-}
-
-::ng-deep .mat-progress-bar-fill::after {
-  background-color: #00823C;
-}
-
-.form-container {
-  flex-basis: 360px;
-  background: #3f4652;
-  color: white;
-  overflow-y: auto;
-}
-
-.form {
-  padding: 32px;
-  padding-top: 0;
-}
-
-.product-type {
-  font-size: 28px;
-  font-weight: bold;
-}
-
-.product-size {
-  font-size: 17px;
-  color: rgb(15, 195, 89);
-  font-weight: bold;
-}
-
-.form-field {
-  margin-bottom: 2px;
-}
-
-.form-field label {
-  font-size: 12px;
-  margin-bottom: 2px;
-}
-
-.choose-design-button {
-  position: absolute;
-  top: 50%;
-  margin-left: -64px;
-  z-index: 999;
-}
-
-.back {
-  height: 64px;
-  display: flex;
-  width: 100%;
-  justify-content: space-between;
-  align-items: center;
-  padding: 0 24px 0 12px;
-  background-color: #eeeeee;
-  position: relative;
-  z-index: 99;
-}
-
-.back button {
-  outline: 0;
-  border: 0;
-}
-
-.selected-tab {
-  position: absolute;
-  z-index: 0;
-  left: 0;
-  top: 0;
-  width: 70px;
-  height: 70px;
-  background-color: #3f4652;
-  transition: transform .3s ease-out;
-  transform: translateY(280px);
-  line-height: 1.4;
-  font-family: Open Sans,Helvetica Neue,Helvetica,Arial,sans-serif;
-  font-size: 14px;
-  text-align: left;
-}
-.selected-tab span {
-  position: absolute;
-  right: 0;
-  top: -7px;
-  width: 7px;
-  height: 7px;
-  background-color: #3f4652;
-}
-.bottom {
-  top: auto !important;
-  bottom: -7px !important;
-}
-.selected-tab span:after {
-  content: '';
-  position: absolute;
-  right: 0;
-  bottom: 0;
-  width: 14px;
-  height: 14px;
-  border-radius: 7px;
-  background-color: rgb(48,52,58);
-}
-.bottom:after {
-  top: 0 !important;
-  bottom: auto !important;
-}
-
-.side-nav {
-  overflow: hidden;
-  position: relative;
-  height: 100%;
-  width: 70px;
-  background-color: transparent !important;
-}
-
-.tab {
-  width: 70px;
-  height: 70px;
-  padding: 16px 0;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: space-around;
-  cursor: pointer;
-  color: #999;
-  box-shadow: none;
-  transition: background-color 0.1s, color 0.1s;
-  text-align: center;
-  text-transform: uppercase;
-}
-.tab:hover {
-  color: rgb(18, 196, 99);
-}
-.tab.selected {
-  color: white;
-}
-.tab-title {
-  font-size: 12px;
-}
-
-#canvas {
-  cursor: default !important;
-}
