--- src/app/designer/dev/designer-dev.component.html
@@ -1,82 +1,0 @@
-<div class="checkout-popup" [class.hidden]="!checkingOut">
-  <router-outlet></router-outlet>
-</div>
-<div class="checkout-background" [class.hidden]="!checkingOut"></div>
-
-<app-designer-view [loading]="loading">
-  <ng-container sidebar>
-    <!-- BEGIN sidebar tabs -->
-    <app-sidebar-tab title="Agents"
-                     icon="contact_mail"
-                     [disabled]="!selectedProduct || !designState.canvasData">
-      <app-agents [user]="selectedAgent"
-                  [agents]="agents"
-                  [loading]="loading"
-                  (changeAgent)="changeAgent($event)"
-                  [(selectedAgent)]="selectedAgent">
-      </app-agents>
-    </app-sidebar-tab>
-    <app-sidebar-tab title="Products"
-                     icon="aspect_ratio">
-      <app-products (changeProduct)="changeProduct($event)"
-                    [selectedProduct]="selectedProduct"></app-products>
-    </app-sidebar-tab>
-    <app-sidebar-tab #designsTab
-                     title="Designs"
-                     icon="view_carousel"
-                     [disabled]="!selectedProduct">
-      <app-designs [size]="'9x6'"
-                   (select)="loadDesign($event)">
-      </app-designs>
-    </app-sidebar-tab>
-    <app-sidebar-tab title="Property"
-                     icon="home"
-                     [disabled]="!selectedProduct || !designState.canvasData">
-      <app-property [agent]="selectedAgent"
-                    [listingId]="listingId"
-                    [(selectedListing)]="selectedListing"
-                    (changeProperty)="changeProperty($event)">
-      </app-property>
-    </app-sidebar-tab>
-    <app-sidebar-tab title="Text"
-                     icon="subject"
-                     [disabled]="!selectedProduct || !designState.canvasData">
-      <app-text [formFields]="designState.textFields"
-                (render)="fabricCanvas.render()"></app-text>
-    </app-sidebar-tab>
-    <app-sidebar-tab title="Colors"
-                     icon="color_lens"
-                     [disabled]="!selectedProduct || !designState.canvasData">
-      <app-colors *ngIf="orderState.brandColors"
-                  [(colors)]="orderState.brandColors">
-      </app-colors>
-    </app-sidebar-tab>
-    <!-- END sidebar tabs -->
-  </ng-container>
-
-  <ng-container designer>
-    <div class="designer-view-options">
-      <button id="front"
-              class="btn"
-              [class.active]="viewSide === 'front'"
-              (click)="setViewSide('front')"
-              [disabled]="!selectedProduct || !designState.canvasData || processing">View Front</button>
-      <button id="back"
-              class="btn"
-              [class.active]="viewSide === 'back'"
-              (click)="setViewSide('back')"
-              [disabled]="!selectedProduct || !designState.canvasData || processing">View Back</button>
-    </div>
-    <!-- canvas -->
-    <app-fabric-canvas #canvas
-                       [loading]="processing"
-                       (click)="clickObject($event)"></app-fabric-canvas>
-    <div class="designer-view-save">
-      <button id="save"
-              class="btn btn-green"
-              [disabled]="!selectedProduct || !designState.canvasData || processing"
-              (click)="saveAndContinue()">Save and Continue</button>
-    </div>
-  </ng-container>
-
-</app-designer-view>
